Added functionality for autogeneration of credentials

+++ src/codes/python/django/nb/ISAD/src/vlabs/utils/generate_secret_key.py
@@ -6,8 +6,8 @@
 #
 # NOTE: This script should be execute once (and only once) after an app
 # is deployed to a new server machine. Even if the source code is
-# modified, a given instance of the lab should continue using the same 
-# secret key. A new secret key should be generated if a new instance of 
+# modified, a given instance of the lab should continue using the same
+# secret key. A new secret key should be generated if a new instance of
 # this app is deployed to a new server.
 #
 # In short, the secret key is unique for each installation of this lab.
@@ -17,7 +17,7 @@
 #
 # Source: http://blog.leosoto.com/2008/04/django-secretkey-generation.html
 #
-# Barun Saha (barun.saha04@gmail.com)
+# Barun Saha (http://barunsaha.me)
 # IIT Kharagpur
 # 06 March 2014
 #
@@ -28,30 +28,33 @@ import stat
 import sys
 
 
-secret = "".join(
-	[random.choice("abcdefghijklmnopqrstuvwxyz0123456789!@#$%^&*(-_=+)") \
-		for i in range(50)]
-)
+def generate_secret():
+	secret = "".join(
+		[random.choice("abcdefghijklmnopqrstuvwxyz0123456789!@#$%^&*(-_=+)") \
+			for i in range(50)]
+	)
 
-#print secret
+	#print secret
 
-file_name = 'secret.txt'
+	file_name = 'secret.txt'
 
-try:
-	
-	#with open(file_name, 'w') as sfile:
-	open(file_name, 'w').write(secret)
-	
-	# Read-only by owner and group
-	os.chmod(file_name, stat.S_IRUSR | stat.S_IRGRP)
-	
-except IOError, ioe:
-	print ioe
-	sys.exit(1)
+	try:
+
+		#with open(file_name, 'w') as sfile:
+		file_exists = os.path.exists(file_name)
+		if not file_exists:
+			open(file_name, 'w').write(secret)
+
+		# Read-only by owner and group
+		os.chmod(file_name, stat.S_IRUSR | stat.S_IRGRP)
+
+	except IOError, ioe:
+		print ioe
+		sys.exit(1)
 
 
 
-#
-# The secret key has been generated. Read it from the settings.py file
-# as open('secret.txt', 'r').read()
-#
+	#
+	# The secret key has been generated. Read it from the settings.py file
+	# as open('secret.txt', 'r').read()
+	#
